refactor in dao classes, valid all id's

+++ server/database/msg_dao.py
@@ -1,7 +1,7 @@
 from bson import ObjectId
 
 from server.database import database
-from server.database.database import DB
+from server.database.database import DB, id_is_valid
 from server.entities.chats.inner_classes.message import Message
 
 msg_collection = DB['messages']
@@ -17,6 +17,9 @@ def save_msg(msg):
 
 
 def get_msg(id):
+    if not id_is_valid(id):
+        return None
+
     json = msg_collection.find_one({'_id': ObjectId(id)})
     if json is None:
         return None
@@ -28,8 +31,14 @@ def get_msg(id):
 
 
 def delete_msg(id):
+    if not id_is_valid(id):
+        return 0
+
     return msg_collection.delete_one({'_id': ObjectId(id)}).deleted_count
 
 
 def is_exists(id):
+    if not id_is_valid(id):
+        return False
+
     return database.is_exist(id, msg_collection)
